Reorganise flows and tasks into separate dirs
