+++ server/src/tasks/mock.py
@@ -1,20 +1,7 @@
-"""This module implements a MapReduce wordcount program as a Flow
-to test a distributed Prefect pipeline. This classic MapReduce program
-was also chosen to demonstrate Prefect's Dask executor, which
-`streamlit-e2e-boilerplate` deploys onto a Kubernetes cluster.
-
-Note: this module is not meant to be an efficient solution to the word
-counting problem. It is only meant to demonstrate distributed workflows
-in Prefect.
-"""
-
 import itertools
 import requests
 
-from prefect import Parameter
 from prefect import task
-from prefect import Flow
-from prefect import flatten
 
 
 @task
@@ -54,17 +41,6 @@ def shuffler(token_tuples):
 def reducer(partition):
     key, value = partition
     return (key, sum(value))
-
-
-with Flow(name='mock_flow') as mock_flow:
-
-    url = Parameter('url', required=True)
-
-    message = download_message(url)
-    lines = split_message(message)
-    token_tuples = mapper.map(lines)
-    partitions = shuffler(flatten(token_tuples))
-    token_counts = reducer.map(partitions)
 
 
 if __name__ == "__main__":
