+++ server/src/tests/test_flows.py
@@ -1,0 +1,7 @@
+import pytest
+
+
+
+
+if __name__ == "__main__":
+    pass
